requirements.txt
django>=6.0
psycopg2
matplotlib
#basemap

djangorestframework
django-cors-headers
django-jquery
django-import-export
django-select2
djangorestframework
django-json-widget
django-tables2
django-filter
django-bootstrap5
martor
markdown
django-filter
django-q2
django-storages[s3]
boto3

redis

pandas 
numpy
fastparquet
plotly
gpxpy

# QR Code generation
qrcode[pil]
Pillow

django-crispy-forms
crispy-bootstrap5
#django-organizations
django-guardian
django-gravatar2
django-prettyjson

django-markdownx
drf-spectacular

markdown
djangorestframework
setuptools

pytest
pytest-django